refactor, dev: бизнес логика вынесена из папки handlers в отдельную папку
